refactor: SessionDetail.tsx のドメインコンポーネントを session/ ディレクトリに分離
ToolCallView, StreamDisplayItemView, StreamOutputView, DiffDropdown,
ToolInputView を個別ファイルに抽出し、SessionDetail.tsx をオーケストレーターとして整理。
