+++ frontend/src/components/SessionDetail.tsx
@@ -3,16 +3,12 @@ import { useParams, useNavigate } from "react-router-dom";
 import Markdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import {
-  Square, RefreshCw, Trash2, ArrowLeft, GitBranch, GitPullRequest, FileDiff, X, FolderOpen,
-  Sparkles, CheckCircle2,
-  ListTodo, Target, RotateCcw, Circle, ImagePlus, SendHorizonal, AlertTriangle, Plus, Slash,
-  Code, Eye,
+  Square, RefreshCw, Trash2, ArrowLeft, GitBranch, GitPullRequest, FolderOpen,
+  Sparkles,
+  ListTodo, Target, RotateCcw, ImagePlus, SendHorizonal, Plus, Slash,
+  Code, Eye, X,
 } from "lucide-react";
 import { Modal } from "./ui/Modal";
-import { CollapsibleText } from "./ui/CollapsibleText";
-import type { StreamEntry, StreamDisplayItem, AskUserQuestionItem } from "../models/stream";
-import { mergeStreamEntries } from "../models/stream";
-import { toolIcon, toolDescription, toolSubDetail, parseTodoInput } from "../models/tool";
 import { Toast } from "./ui/Toast";
 import { useAutoScroll } from "../hooks/useAutoScroll";
 import type { Session } from "../types/session";
@@ -20,564 +16,8 @@ import { api, type DiffFile } from "../api/client";
 import { StatusDot } from "./StatusBadge";
 import { setActiveSessionName } from "../activeSession";
 import { useWebSocket } from "../hooks/useWebSocket";
-
-const roleStyles: Record<string, { bg: string; label: string; text: string }> = {
-  user: { bg: "bg-blue-50", label: "User", text: "text-blue-700" },
-  assistant: { bg: "bg-gray-50", label: "Assistant", text: "text-green-700" },
-};
-
-
-
-
-
-function ToolInputView({ input }: { input: unknown }) {
-  if (input && typeof input === "object" && !Array.isArray(input)) {
-    const entries = Object.entries(input as Record<string, unknown>);
-    return (
-      <div className="space-y-2">
-        {entries.map(([key, value]) => {
-          const str = typeof value === "string" ? value : JSON.stringify(value, null, 2);
-          const isMultiline = str.includes("\n");
-          return (
-            <div key={key}>
-              <span className="text-gray-400 text-[10px] uppercase tracking-wide">{key}</span>
-              {isMultiline ? (
-                <pre className="text-gray-700 text-xs mt-0.5 bg-gray-50 border border-gray-200 rounded p-2 overflow-x-auto whitespace-pre-wrap">{str}</pre>
-              ) : (
-                <div className="text-gray-700 text-xs mt-0.5 font-mono">{str}</div>
-              )}
-            </div>
-          );
-        })}
-      </div>
-    );
-  }
-  const str = typeof input === "string" ? input : JSON.stringify(input, null, 2);
-  return <pre className="text-gray-600 text-xs overflow-x-auto whitespace-pre-wrap">{str}</pre>;
-}
-
-function AskUserQuestionCallView({ item, onAnswer }: { item: Extract<StreamDisplayItem, { kind: "tool_call" }>; onAnswer?: (text: string) => void }) {
-  const [expanded, setExpanded] = useState(true);
-  const inp = item.input as { questions?: AskUserQuestionItem[] } | undefined;
-  const Icon = toolIcon(item.name);
-  const desc = toolDescription(item.name, item.input);
-
-  return (
-    <div className="border border-orange-200 rounded-lg overflow-hidden bg-orange-50">
-      <button
-        onClick={() => setExpanded(!expanded)}
-        className="w-full text-left px-2.5 py-1.5 hover:bg-orange-100 transition-colors"
-      >
-        <div className="flex items-center gap-2">
-          <Icon className="w-3.5 h-3.5 text-orange-500 shrink-0" />
-          <span className="font-medium text-xs text-orange-800">{item.name}</span>
-          {desc && (
-            <span className="text-xs text-orange-600 truncate min-w-0">{desc}</span>
-          )}
-        </div>
-      </button>
-      {expanded && inp?.questions && (
-        <div className="px-3 pb-3 space-y-3">
-          {inp.questions.map((q, qi) => (
-            <div key={qi}>
-              {q.header && (
-                <span className="inline-block px-1.5 py-0.5 text-[10px] font-medium bg-orange-200 text-orange-800 rounded mb-1">
-                  {q.header}
-                </span>
-              )}
-              <p className="text-sm text-gray-800 mb-2">{q.question}</p>
-              <div className="flex flex-wrap gap-2">
-                {q.options.map((opt, oi) => (
-                  <button
-                    key={oi}
-                    onClick={() => onAnswer?.(opt.label)}
-                    className="px-3 py-1.5 text-xs rounded-lg border bg-white border-gray-300 text-gray-700 hover:bg-blue-50 hover:border-blue-300 hover:text-blue-700 transition-colors"
-                    title={opt.description}
-                  >
-                    {opt.label}
-                  </button>
-                ))}
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function formatTimeout(seconds: number): string {
-  if (seconds >= 60) {
-    const min = Math.floor(seconds / 60);
-    return `${min}分`;
-  }
-  return `${seconds}秒`;
-}
-
-function EscalateCallView({ item, sessionId, escalationId, timedOut, timeoutSeconds, onResponded }: {
-  item: Extract<StreamDisplayItem, { kind: "tool_call" }>;
-  sessionId?: string;
-  escalationId?: string;
-  timedOut?: boolean;
-  timeoutSeconds?: number;
-  onResponded?: () => void;
-}) {
-  const [expanded, setExpanded] = useState(true);
-  const [response, setResponse] = useState("");
-  const [sending, setSending] = useState(false);
-  const inp = item.input as { message?: string } | undefined;
-  const isResolved = item.result !== undefined;
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!sessionId || !response.trim() || sending) return;
-
-    if (!escalationId) return;
-    const escId = escalationId;
-    setSending(true);
-    try {
-      await api.respondEscalation(sessionId, escId, response.trim());
-      setResponse("");
-      onResponded?.();
-    } catch {
-      // ignore errors
-    } finally {
-      setSending(false);
-    }
-  };
-
-  return (
-    <div className="border border-red-200 rounded-lg overflow-hidden bg-red-50">
-      <button
-        onClick={() => setExpanded(!expanded)}
-        className="w-full text-left px-2.5 py-1.5 hover:bg-red-100 transition-colors"
-      >
-        <div className="flex items-center gap-2">
-          <AlertTriangle className="w-3.5 h-3.5 text-red-500 shrink-0" />
-          <span className="font-medium text-xs text-red-800">Escalation</span>
-          {timeoutSeconds && !isResolved && !timedOut && (
-            <span className="text-xs text-gray-500">
-              (タイムアウト: {formatTimeout(timeoutSeconds)})
-            </span>
-          )}
-          {timedOut && !isResolved && (
-            <span className="text-xs text-amber-600 bg-amber-50 border border-amber-200 rounded px-1.5 py-0.5">
-              タイムアウト - エージェントは自動続行しました
-            </span>
-          )}
-          {isResolved && <CheckCircle2 className="w-3 h-3 text-green-500 ml-auto shrink-0" />}
-        </div>
-      </button>
-      {expanded && (
-        <div className="px-3 pb-3 space-y-2">
-          <p className="text-sm text-gray-800">{inp?.message}</p>
-          {isResolved ? (
-            <div className="text-xs text-gray-500 bg-white border border-gray-200 rounded px-2 py-1.5">
-              <span className="text-gray-400">Response: </span>
-              {item.result}
-            </div>
-          ) : timedOut ? (
-            <div className="text-xs text-amber-700 bg-amber-50 border border-amber-200 rounded px-2 py-1.5">
-              タイムアウト - エージェントは自動続行しました
-            </div>
-          ) : (
-            <form onSubmit={handleSubmit} className="flex gap-2">
-              <input
-                type="text"
-                value={response}
-                onChange={(e) => setResponse(e.target.value)}
-                placeholder="Enter your response..."
-                className="flex-1 border border-gray-300 rounded px-2 py-1.5 text-xs"
-                disabled={sending}
-              />
-              <button
-                type="submit"
-                disabled={sending || !response.trim()}
-                className="px-3 py-1.5 text-xs bg-red-600 text-white rounded hover:bg-red-700 disabled:opacity-50"
-              >
-                {sending ? "..." : "Respond"}
-              </button>
-            </form>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function ToolCallView({ item, onAnswer, sessionId, escalationId, escalationTimedOut, escalationTimeoutSeconds, onEscalationResponded }: {
-  item: Extract<StreamDisplayItem, { kind: "tool_call" }>;
-  onAnswer?: (text: string) => void;
-  sessionId?: string;
-  escalationId?: string;
-  escalationTimedOut?: boolean;
-  escalationTimeoutSeconds?: number;
-  onEscalationResponded?: () => void;
-}) {
-  if (item.name === "TodoWrite") {
-    return <TodoCallView item={item} />;
-  }
-  if (item.name === "AskUserQuestion") {
-    return <AskUserQuestionCallView item={item} onAnswer={onAnswer} />;
-  }
-  if (item.name === "mcp__agmux__escalate") {
-    return <EscalateCallView item={item} sessionId={sessionId} escalationId={escalationId} timedOut={escalationTimedOut} timeoutSeconds={escalationTimeoutSeconds} onResponded={onEscalationResponded} />;
-  }
-
-  const [open, setOpen] = useState(false);
-  const [childrenExpanded, setChildrenExpanded] = useState(false);
-  const Icon = toolIcon(item.name);
-  const desc = toolDescription(item.name, item.input);
-  const subDetail = toolSubDetail(item.name, item.input);
-  const done = item.result !== undefined;
-  const hasChildren = item.children && item.children.length > 0;
-  const CHILD_DISPLAY_LIMIT = 5;
-
-  return (
-    <>
-      <button
-        onClick={() => setOpen(true)}
-        className="w-full text-left border border-gray-200 rounded-lg overflow-hidden px-2.5 py-1.5 bg-gray-50 hover:bg-gray-100 transition-colors"
-      >
-        <div className="flex items-center gap-2">
-          <Icon className="w-3.5 h-3.5 text-gray-500 shrink-0" />
-          <span className="font-medium text-xs text-gray-800">{item.name}</span>
-          {desc && (
-            <span className="text-xs text-gray-500 truncate min-w-0">{desc}</span>
-          )}
-          {done && <CheckCircle2 className="w-3 h-3 text-green-500 ml-auto shrink-0" />}
-        </div>
-        {subDetail && (
-          <div className="mt-0.5 ml-[22px] font-mono text-[11px] text-gray-400 truncate">{subDetail}</div>
-        )}
-      </button>
-      <Modal
-        open={open}
-        onClose={() => setOpen(false)}
-        title={<span className="flex items-center gap-2"><Icon className="w-4 h-4 text-gray-500" />{item.name} {desc && <span className="text-gray-400 font-normal">{desc}</span>}</span>}
-      >
-        <div className="space-y-3">
-          <ToolInputView input={item.input} />
-          {item.resultImages && item.resultImages.length > 0 && (
-            <div>
-              <span className="text-gray-400 text-[10px] uppercase tracking-wide">Images</span>
-              <div className="mt-1 flex flex-wrap gap-2">
-                {item.resultImages.map((img, i) => (
-                  <img
-                    key={i}
-                    src={`data:${img.mediaType};base64,${img.data}`}
-                    alt={`result-${i}`}
-                    className="max-w-full max-h-96 rounded border border-gray-200"
-                  />
-                ))}
-              </div>
-            </div>
-          )}
-          {done && item.result && (
-            <div>
-              <span className="text-gray-400 text-[10px] uppercase tracking-wide">Output</span>
-              <pre className="text-gray-600 text-xs overflow-x-auto whitespace-pre-wrap mt-0.5">{item.result.slice(0, 2000)}</pre>
-            </div>
-          )}
-        </div>
-      </Modal>
-      {hasChildren && (() => {
-        const allChildren = item.children!;
-        const totalCount = allChildren.length;
-        const hasHidden = totalCount > CHILD_DISPLAY_LIMIT && !childrenExpanded;
-        const hiddenCount = totalCount - CHILD_DISPLAY_LIMIT;
-        const visibleChildren = hasHidden
-          ? allChildren.slice(totalCount - CHILD_DISPLAY_LIMIT)
-          : allChildren;
-        return (
-          <div className="ml-4 border-l-2 border-indigo-200 pl-3 space-y-1.5">
-            {hasHidden && (
-              <button
-                onClick={() => setChildrenExpanded(true)}
-                className="text-xs text-blue-500 hover:text-blue-700 py-0.5"
-              >
-                他 {hiddenCount} 件を表示
-              </button>
-            )}
-            {childrenExpanded && totalCount > CHILD_DISPLAY_LIMIT && (
-              <button
-                onClick={() => setChildrenExpanded(false)}
-                className="text-xs text-blue-500 hover:text-blue-700 py-0.5"
-              >
-                折りたたむ
-              </button>
-            )}
-            {visibleChildren.map((child, i) => (
-              <div key={hasHidden ? i + hiddenCount : i}>
-                {child.kind === "tool_call" ? (
-                  <ToolCallView item={child} />
-                ) : child.kind === "text" ? (
-                  <CollapsibleText text={child.text} />
-                ) : null}
-              </div>
-            ))}
-          </div>
-        );
-      })()}
-    </>
-  );
-}
-
-function SystemEventView({ item }: { item: Extract<StreamDisplayItem, { kind: "system_event" }> }) {
-  return (
-    <div className="flex items-center gap-2 py-1 px-3 text-xs text-gray-500 border-y border-dashed border-gray-200">
-      <span className="text-gray-400">⟡</span>
-      <span>{item.label}</span>
-      {item.detail && <span className="text-gray-400">{item.detail}</span>}
-    </div>
-  );
-}
-
-
-function StreamDisplayItemView({ item, onAnswer, sessionId, escalationId, escalationTimedOut, escalationTimeoutSeconds, onEscalationResponded }: {
-  item: StreamDisplayItem;
-  onAnswer?: (text: string) => void;
-  sessionId?: string;
-  escalationId?: string;
-  escalationTimedOut?: boolean;
-  escalationTimeoutSeconds?: number;
-  onEscalationResponded?: () => void;
-}) {
-  if (item.kind === "text") {
-    return <CollapsibleText text={item.text} />;
-  }
-  if (item.kind === "image") {
-    return (
-      <img
-        src={`data:${item.mediaType};base64,${item.data}`}
-        alt="attached"
-        className="max-w-xs max-h-48 rounded border border-gray-200"
-      />
-    );
-  }
-  if (item.kind === "tool_call") {
-    return <ToolCallView item={item} onAnswer={onAnswer} sessionId={sessionId} escalationId={escalationId} escalationTimedOut={escalationTimedOut} escalationTimeoutSeconds={escalationTimeoutSeconds} onEscalationResponded={onEscalationResponded} />;
-  }
-  if (item.kind === "system_event") {
-    return <SystemEventView item={item} />;
-  }
-  return null;
-}
-
-function TodoCallView({ item }: { item: Extract<StreamDisplayItem, { kind: "tool_call" }> }) {
-  const [open, setOpen] = useState(false);
-  const todos = parseTodoInput(item.input);
-  if (!todos || todos.length === 0) return null;
-
-  const completed = todos.filter(t => t.status === "completed").length;
-
-  return (
-    <>
-      <button
-        onClick={() => setOpen(true)}
-        className="w-full text-left border border-gray-200 rounded-lg overflow-hidden px-2.5 py-1.5 bg-gray-50 hover:bg-gray-100 transition-colors"
-      >
-        <div className="flex items-center gap-2 mb-1.5">
-          <ListTodo className="w-3.5 h-3.5 text-gray-400" />
-          <span className="font-medium text-xs text-gray-800">Tasks</span>
-          <span className="text-[10px] text-gray-400 ml-auto">{completed}/{todos.length}</span>
-        </div>
-        <div className="space-y-0.5">
-          {todos.map((todo, i) => (
-            <div key={i} className="flex items-start gap-1.5 text-xs">
-              {todo.status === "completed" ? (
-                <CheckCircle2 className="w-3.5 h-3.5 text-green-500 shrink-0 mt-0.5" />
-              ) : todo.status === "in_progress" ? (
-                <Circle className="w-3.5 h-3.5 text-blue-500 fill-blue-500 shrink-0 mt-0.5" />
-              ) : (
-                <Circle className="w-3.5 h-3.5 text-gray-300 shrink-0 mt-0.5" />
-              )}
-              <span className={todo.status === "completed" ? "text-gray-400 line-through" : "text-gray-700"}>
-                {todo.content}
-              </span>
-            </div>
-          ))}
-        </div>
-      </button>
-      <Modal
-        open={open}
-        onClose={() => setOpen(false)}
-        title={<span className="flex items-center gap-2"><ListTodo className="w-4 h-4 text-gray-500" />TodoWrite</span>}
-      >
-        <div className="space-y-3">
-          <ToolInputView input={item.input} />
-          {item.result !== undefined && (
-            <div>
-              <span className="text-gray-400 text-[10px] uppercase tracking-wide">Output</span>
-              <pre className="text-gray-600 text-xs overflow-x-auto whitespace-pre-wrap mt-0.5">{item.result!.slice(0, 2000)}</pre>
-            </div>
-          )}
-        </div>
-      </Modal>
-    </>
-  );
-}
-
-// --- Shared components ---
-
-type StreamViewMode = "markdown" | "json";
-
-function StreamOutputView({ lines, className, onAnswer, sessionId, escalationId, escalationTimedOut, escalationTimeoutSeconds, onEscalationResponded }: {
-  lines: unknown[];
-  className?: string;
-  onAnswer?: (text: string) => void;
-  sessionId?: string;
-  escalationId?: string;
-  escalationTimedOut?: boolean;
-  escalationTimeoutSeconds?: number;
-  onEscalationResponded?: () => void;
-}) {
-  const { ref, onScroll } = useAutoScroll(lines);
-  const [viewMode, setViewMode] = useState<StreamViewMode>("markdown");
-
-  const entries = lines
-    .map((line) => line as StreamEntry)
-    .filter((e) => e.type === "user" || e.type === "assistant" || e.type === "system");
-
-  const groups = mergeStreamEntries(entries);
-
-  return (
-    <div className={`flex flex-col ${className || ""}`}>
-      <div className="flex justify-end mb-1 shrink-0">
-        <button
-          onClick={() => setViewMode(viewMode === "markdown" ? "json" : "markdown")}
-          className="text-xs text-gray-500 hover:text-gray-700 px-2 py-1"
-        >
-          {viewMode === "markdown" ? "JSON" : "Markdown"}
-        </button>
-      </div>
-      <div ref={ref} onScroll={onScroll} className="bg-white border border-gray-200 rounded-lg p-3 text-sm flex-1 min-h-0 overflow-y-auto space-y-3">
-        {viewMode === "json" ? (
-          lines.length === 0 ? (
-            <p className="text-gray-400">No stream output yet</p>
-          ) : (
-            lines.map((line, i) => (
-              <pre key={i} className="text-gray-600 text-xs whitespace-pre-wrap">
-                {JSON.stringify(line, null, 2)}
-              </pre>
-            ))
-          )
-        ) : groups.length === 0 ? (
-          <p className="text-gray-400">No stream output yet</p>
-        ) : (
-          groups.map((group, i) => {
-            if (group.role === "system") {
-              return (
-                <div key={i}>
-                  {group.items.map((item, j) => (
-                    <StreamDisplayItemView key={j} item={item} onAnswer={onAnswer} sessionId={sessionId} escalationId={escalationId} escalationTimedOut={escalationTimedOut} escalationTimeoutSeconds={escalationTimeoutSeconds} onEscalationResponded={onEscalationResponded} />
-                  ))}
-                </div>
-              );
-            }
-            const style = roleStyles[group.role] || roleStyles.assistant;
-            return (
-              <div key={i} className={`rounded-lg p-3 ${style.bg}`}>
-                <div className="flex items-center gap-2 mb-1">
-                  <span className={`font-semibold text-xs ${style.text}`}>
-                    {style.label}
-                  </span>
-                </div>
-                <div className="text-gray-800 break-words text-xs space-y-2">
-                  {group.items.map((item, j) => (
-                    <StreamDisplayItemView key={j} item={item} onAnswer={onAnswer} sessionId={sessionId} escalationId={escalationId} escalationTimedOut={escalationTimedOut} escalationTimeoutSeconds={escalationTimeoutSeconds} onEscalationResponded={onEscalationResponded} />
-                  ))}
-                </div>
-              </div>
-            );
-          })
-        )}
-      </div>
-    </div>
-  );
-}
-
-const statusBadgeColor: Record<string, string> = {
-  M: "bg-yellow-100 text-yellow-800",
-  A: "bg-green-100 text-green-800",
-  D: "bg-red-100 text-red-800",
-  R: "bg-blue-100 text-blue-800",
-  "?": "bg-gray-100 text-gray-600",
-};
-
-function DiffDropdown({ files }: { files: DiffFile[] }) {
-  const [open, setOpen] = useState(false);
-  const [expanded, setExpanded] = useState<Set<string>>(new Set());
-
-  if (files.length === 0) return null;
-
-  const toggle = (path: string) => {
-    setExpanded((prev) => {
-      const next = new Set(prev);
-      if (next.has(path)) next.delete(path);
-      else next.add(path);
-      return next;
-    });
-  };
-
-  return (
-    <>
-      <button
-        onClick={() => setOpen(true)}
-        className="inline-flex items-center gap-1 text-xs px-2 py-0.5 rounded-full bg-orange-50 text-orange-700 hover:bg-orange-100"
-        title="Changes"
-      >
-        <FileDiff className="w-3 h-3" />
-        {files.length}
-      </button>
-      <Modal
-        open={open}
-        onClose={() => setOpen(false)}
-        title={<span className="flex items-center gap-2"><FileDiff className="w-4 h-4 text-orange-500" />Changes ({files.length} files)</span>}
-      >
-        <div className="border border-gray-200 rounded-lg overflow-hidden">
-          {files.map((file) => (
-            <div key={file.path} className="border-b border-gray-100 last:border-b-0">
-              <button
-                onClick={() => toggle(file.path)}
-                className="w-full flex items-center gap-2 px-3 py-1.5 text-xs hover:bg-gray-50 text-left min-w-0"
-              >
-                <span className={`px-1.5 py-0.5 rounded font-mono text-[10px] font-bold shrink-0 ${statusBadgeColor[file.status] || "bg-gray-100 text-gray-600"}`}>
-                  {file.status}
-                </span>
-                <span className="font-mono text-gray-700 truncate min-w-0">{file.path}</span>
-                {file.diff && (
-                  <span className="ml-auto text-gray-400 shrink-0">{expanded.has(file.path) ? "▼" : "▶"}</span>
-                )}
-              </button>
-              {expanded.has(file.path) && file.diff && (
-                <pre className="bg-gray-50 text-gray-800 text-xs p-3 overflow-x-auto whitespace-pre font-mono border-t border-gray-200">
-                  {file.diff.split("\n").map((line, i) => (
-                    <span
-                      key={i}
-                      className={
-                        line.startsWith("+") && !line.startsWith("+++")
-                          ? "text-green-700 bg-green-50"
-                          : line.startsWith("-") && !line.startsWith("---")
-                            ? "text-red-700 bg-red-50"
-                            : line.startsWith("@@")
-                              ? "text-blue-600"
-                              : ""
-                      }
-                    >
-                      {line}
-                      {"\n"}
-                    </span>
-                  ))}
-                </pre>
-              )}
-            </div>
-          ))}
-        </div>
-      </Modal>
-    </>
-  );
-}
+import { StreamOutputView } from "./session/StreamOutputView";
+import { DiffDropdown } from "./session/DiffDropdown";
 
 export function SessionDetail() {
   const { id: sessionId } = useParams<{ id: string }>();
